+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this branch will be documented in this file.
 
 
+## 1.4.0 - Released
+
+### Added
+
+- Added RTL (Right-to-Left) support for Arabic like languages, allowing the app to display correctly in these languages.
+- Introduced a `ContentDialog` to display scan results on the `Repair` page #44.
+- Added support for Hebrew language (`he-il`), contributed by @Y-PLONI #36.
+
+### Fixes
+
+- Fixed some wrong translations in the `zh-Hans` language, contributed by @wumingshiali #37.
+
+### Changes
+
+- Changed the restore point creation process to use the default System Restore `SystemPropertiesProtection` #41.
+- Enhance telemetry disabling and temp file cleanup.
+
 ## 1.3.2 - Released
 
 ### Added